correcting inaccuracies and creating a single main
